+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Update Vuelidate !3889 @SaiCode
 - Rework working group rest controller !3879 @chriswalg
 - Rework WallRestController !3880 @chriswalg
+- Removed jquery !3884 @chriswalg
 - Moved basket store to pinia !3687 @chriswalg
 - Remove jquery-ui and some other jquery extentions !3883 @chriswalg
 - Rework foodsaver rest controller !3892 @chriswalg